Move COmniTransactionDB into seperate file
